Move Tensorflow method to vendored.py.
This moves _create_file_writer_generic_type to vendored.py.
