+++ sticker-graph/sticker_graph/vendored.py
@@ -92,3 +92,32 @@ def stack_bidirectional_dynamic_rnn(cells_fw,
             states_bw.append(state_bw)
 
     return prev_layer, tuple(states_fw), tuple(states_bw)
+
+
+def _create_file_writer_generic_type(logdir,
+                                    name="logdir",
+                                    max_queue=None,
+                                    flush_millis=None,
+                                    filename_suffix=None):
+    """
+    This method mirrors `tensorflow.contrib.summary.create_file_writer`. Unlike
+    `summary.create_file_writer`, this method accepts a placeholder as `logdir`.
+    """
+    from tensorflow.python.ops.summary_ops_v2 import _make_summary_writer
+    from tensorflow.python.ops.gen_summary_ops import create_summary_file_writer
+
+    logdir = tf.convert_to_tensor(logdir)
+    with tf.device("cpu:0"):
+        if max_queue is None:
+            max_queue = tf.constant(10)
+        if flush_millis is None:
+            flush_millis = tf.constant(2 * 60 * 1000)
+        if filename_suffix is None:
+            filename_suffix = tf.constant(".v2")
+        return _make_summary_writer(
+            name,
+            create_summary_file_writer,
+            logdir=logdir,
+            max_queue=max_queue,
+            flush_millis=flush_millis,
+            filename_suffix=filename_suffix)
